Move CA1052 fixer to ApiGuidelinesAnalyzer. No functional change.
